backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# MongoDB async driver
motor==3.7.0
pymongo==4.10.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20

# Settings
pydantic==2.10.4
pydantic-settings==2.7.0

# LLM SDKs
openai==1.59.9
google-generativeai==0.8.3
anthropic==0.43.0

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
